refactor: 94- raname registrationState to input-state
