--- src/app/state/login.ts
@@ -1,12 +1,0 @@
-import { type LoginState } from '../types';
-
-export const loginState: LoginState = {
-  email: {
-    error: undefined,
-    value: '',
-  },
-  password: {
-    error: undefined,
-    value: '',
-  },
-};
