+++ src/shared/ui-config/credential-inputs.ts
@@ -1,6 +1,6 @@
 import type { Options } from '../../utils/create-elements/types';
 
-import { registrationState } from '../../app/state/registration';
+import { registrationState } from '../../app/state/input-state';
 import { inputValidation } from '../../utils/validation/input-validation';
 import { ERROR_MESSAGES, REGEX } from '../constants';
 import { INPUT } from '../styles';
